// src/pages/weather/type.ts
export interface WeatherData {
    coord: {
        lon: number;
        lat: number;
    };
    weather: {
        id: number;
        main: string;
        description: string;
        icon: string;
    }[];
    base: string;
    main: {
        temp: number;
        feels_like: number;
        temp_min: number;
        temp_max: number;
        pressure: number;
        humidity: number;

    };
    visibility: number;
    wind: {
        speed: number;
        deg: number;
        gust?: number;
    };
    rain?: {
        "1h"?: number;
        "3h"?: number; 
    };
    clouds: {
    all: number;
    };
    dt: number;
    sys: {
    country: string;
    sunrise: number;
    sunset: number;
    };
    timezone: number;
    id: number;
    name: string;
    cod: number;
}